Alex_Auto_V5_macOS_Updated_Final_Updated_v5_repacked/requirements.txt
Pillow>=8.0
argparse
beautifulsoup4>=4.9.3
colorama
docker>=4.0
gtts
numpy>=1.19
openai>=0.9.0
orjson
python-dotenv
redis>=3.5
regex
requests
subprocess
tiktoken
torch>=1.7
transformers>=4.0
yaml